+++ R/zzz.R
@@ -10,45 +10,14 @@ rob_url <- function(...) paste0(rob_baseurl(), ...)
 rob_get <- function(q_url, token, ...) {
     token <- paste0("Token ", token)
     names(token) <- "Authorization"
-    print(
     httr::GET(q_url, httr::add_headers(token), ...)
-  )
-  httr::GET(q_url, httr::add_headers(token), ...)
 }
 
-rob_res <- function(q_url, token = rob_getsecret(), ...) {
+rob_res <- function(q_url, token, ...) {
     con <- rob_get(q_url, token, ...)
     # check status
     httr::stop_for_status(con)
     # parsed
-    print(httr::content(con, "raw", ...))
-    httr::content(con, "parsed", ...)
+    httr::content(con, "parsed")
 }
 
-rob_getsecret <- function() {
-    val <- Sys.getenv("OWLBOT_TOKEN")
-    if (identical(val, "")) {
-        message("
-            `OWLBOT_TOKEN` env var has not been set yet.
-            A token is required to use the OWLBOT API, see
-            https://owlbot.info/
-        ")
-        rob_set_token()
-        val <- rob_getsecret()
-    }
-    val
-}
-
-rob_set_token <- function(token = NULL) {
-    if (is.null(token))
-        token <- readline("Enter your token without quotes: ")
-    if (identical(token, "")) {
-        stop("No token has been provided.")
-    } else {
-        Sys.setenv(OWLBOT_TOKEN = token)
-        cat("Authentication token stored for the session.\n")
-    }
-    invisible(NULL)
-}
-
-forget_token <- function() Sys.unsetenv("OWLBOT_TOKEN")